added addUser at fake login
