+++ app/src/androidTest/java/ch/epfl/sdp/blindwar/AppStatisticsTest.kt
@@ -15,7 +15,7 @@ class AppStatisticsTest {
 
     @Test
     fun getCorrectInitial() {
-        assert(testStats.correctArray.contentEquals(IntArray(modeSize)))
+        assert(testStats.correctArray.toIntArray().contentEquals(IntArray(modeSize)))
     }
 
     @Test
@@ -31,7 +31,7 @@ class AppStatisticsTest {
 
     @Test
     fun getWrongInitial() {
-        assert(testStats.wrongArray.contentEquals(IntArray(modeSize)))
+        assert(testStats.wrongArray.toIntArray().contentEquals(IntArray(modeSize)))
     }
 
     @Test
