+++ app/src/androidTest/java/ch/epfl/sdp/blindwar/UserTest.kt
@@ -12,7 +12,7 @@ class UserTest {
     private val screenName = "Screen Name"
     private val screenName2 = "Screen Name2"
 
-    private var testUser = User(firstName, lastName, email, screenName, null, AppStatistics())
+    private var testUser = User(firstName, lastName, email, screenName, AppStatistics())
 
     @Test
     fun getFirstName() {
@@ -39,7 +39,7 @@ class UserTest {
         testUser.screenName = screenName2
         assertEquals(testUser.screenName, screenName2)
     }
-
+    /*
     @Test
     fun getProfilePicture() {
         assertNull(testUser.profilePicture)
@@ -48,7 +48,7 @@ class UserTest {
     @Test
     fun setProfilePicture() {
         assertNull(testUser.profilePicture)
-    }
+    }*/
 
     @Test
     fun getUserStatistics() {
