+++ app/src/main/java/ch/epfl/sdp/blindwar/LoginActivity.kt
@@ -4,6 +4,7 @@ import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.view.View
+import ch.epfl.sdp.blindwar.database.UserDatabase
 
 class LoginActivity : AppCompatActivity() {
 
@@ -14,6 +15,11 @@ class LoginActivity : AppCompatActivity() {
     }
 
     fun loginButton(view: View) {
+        val database = UserDatabase()
+        val appStatistics = AppStatistics()
+        val user = User("Jojo", "Joestar", "@", "JOJO",
+            appStatistics)
+        database.addUser(user)
         startActivity(Intent(this, MainMenuActivity::class.java))
     }
 }
