+++ app/src/main/java/ch/epfl/sdp/blindwar/MainMenuActivity.kt
@@ -8,6 +8,8 @@ import ch.epfl.sdp.blindwar.database.UserDatabase
 import ch.epfl.sdp.blindwar.tutorial.TutorialActivity
 
 class MainMenuActivity : AppCompatActivity() {
+
+    private val database = UserDatabase()
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main_menu)
@@ -21,7 +23,6 @@ class MainMenuActivity : AppCompatActivity() {
 
     // Called when the user taps the Tutorial button
     fun tutorialButton(view: View) {
-        val database = UserDatabase()
         database.setElo("JOJO", 1100)
         val intent = Intent(this, TutorialActivity::class.java)
         startActivity(intent)
@@ -30,18 +31,13 @@ class MainMenuActivity : AppCompatActivity() {
     // Called when the user taps the Solo button
     fun profileButton(view: View) {
         val intent = Intent(this, ProfileActivity::class.java)
-        /*
-        val database = UserDatabase()
-        val appStatistics = AppStatistics()
-        val user = User("Jojo", "Joestar", "@", "JOJO",
-            appStatistics)
-        database.addUser(user)*/
         startActivity(intent)
     }
 
     // Called when the user taps the Solo button
     fun logoutButton(view: View) {
         val intent = Intent(this, LoginActivity::class.java)
+        database.setElo("JOJO", 0)
         startActivity(intent)
     }
