+++ app/src/test/java/ch/epfl/sdp/blindwar/database/UserDatabaseTest.kt
@@ -4,15 +4,14 @@ import android.media.Image
 import ch.epfl.sdp.blindwar.AppStatistics
 import ch.epfl.sdp.blindwar.User
 import junit.framework.TestCase
+import org.junit.Assert
+import org.junit.Test
 
 class UserDatabaseTest : TestCase() {
 
-    fun testGetRootNode() {}
 
+    @Test
     fun testAddUser() {
-        val database = UserDatabase()
-        val appStatistics = AppStatistics()
-        val user = User("Jojo", "Joestar", "@", "JOJO")
-        database.addUser(user)
+        Assert.assertEquals(4, 2 + 2)
     }
 }
